@@ -1,8 +1,8 @@
 after :roles, :authors do
   puts('creating users')
-  scott = User.find_or_create_by_email!('scott@copper.is', name:'eScott', accept_terms:true)
-  scott.roles << Role.find_by_name('fan')
-  scott.roles << Role.find_by_name('admin')
+  scott = User.find_or_create_by!(email:'scott@copper.is', name:'eScott', accept_terms:true)
+  scott.roles << Role.find_by(name:'fan')
+  scott.roles << Role.find_by(name:'admin')
   # TODO: Email should be the unique witin the system
 end
 
