@@ -14,7 +14,7 @@ describe TipsController do
     it 'tips pages' do
       proc do
         post_with @me, :create, tip:{url:'http://twitter.com/ableton'}
-        @page = Page.find_by_url('http://twitter.com/ableton')
+        @page = Page.find_by(url:'http://twitter.com/ableton')
         @page.tips.size.should eq(1)
         @me.reload
         @me.current_order.tips.first.page.should eq(@page)
@@ -23,7 +23,7 @@ describe TipsController do
 
     it 'finds authors of tipped pages' do
       post_with @me, :create, tip:{url:'https://twitter.com/ableton'}
-      @page = Page.find_by_url('https://twitter.com/ableton')
+      @page = Page.find_by(url:'https://twitter.com/ableton')
       @page.author.should_not be_nil
       @page.author.username.should eq('ableton')
     end
@@ -31,13 +31,13 @@ describe TipsController do
     it 'messages wanted authors' do
       post_with @me, :create, tip:{url:'http://twitter.com/ableton'}
       @me.current_order.tips.first.destroy
-      @page = Page.find_by_url('http://twitter.com/ableton')
+      @page = Page.find_by(url:'http://twitter.com/ableton')
     end
 
     it 'assign tips to known authors' do
       @her_author = create!(:author_phony)
       post_with @me, :create, tip:{url:"http://example.com/#{@her_author.username}"}
-      @page = Page.find_by_url("http://example.com/#{@her_author.username}")
+      @page = Page.find_by(url:"http://example.com/#{@her_author.username}")
       @page.author.should eq(@her_author)
     end
   end
