@@ -5,9 +5,9 @@ class Check < ActiveRecord::Base
   has_many :tips, :as => :royalties
   validates :user, presence:true
 
-  scope :earned, where("check_state = ?", 'earned')
-  scope :paid, where("check_state = ?", 'paid')
-  scope :cashed, where("check_state = ?", 'cashed')
+  scope :earned, -> { where("check_state = ?", 'earned') }
+  scope :paid,   -> { where("check_state = ?", 'paid') }
+  scope :cashed, -> { where("check_state = ?", 'cashed') }
 
   state_machine :check_state, :initial => :earned do
     event :deliver do
