@@ -1,16 +1,18 @@
 class Order < ActiveRecord::Base
   include Enqueueable
   include OrderMessages
+  has_paper_trail
+
   has_many :tips
   belongs_to :user,  touch:true
-  has_paper_trail
+
   validates :user, presence:true
   validates_associated :user
 
-  scope :current, where('order_state = ?', 'current')
-  scope :unpaid, where('order_state = ?', 'unpaid')
-  scope :denied, where('order_state = ?', 'denied')
-  scope :paid, where('order_state = ?', 'paid')
+  scope :current, -> { where('order_state = ?', 'current') }
+  scope :unpaid,  -> { where('order_state = ?', 'unpaid') }
+  scope :denied,  -> { where('order_state = ?', 'denied') }
+  scope :paid,    -> { where('order_state = ?', 'paid') }
 
   state_machine :order_state, :initial => :current do
     event :process do
@@ -89,4 +91,5 @@ class Order < ActiveRecord::Base
   def total_in_dollars
     sprintf('%.2f', Float(self.total)/100)
   end
+
 end
