@@ -8,7 +8,8 @@ Copper::Application.configure do
   config.active_support.deprecation = :notify
 
 
-  config.assets.js_compressor  = :uglifier
+  config.eager_load = true
+  config.assets.js_compressor = :uglifier
   config.assets.css_compressor = :yui
   config.assets.compress = true
   config.assets.compile = true
